+++ lib/ts/recipe/dashboard/api/userdetails/userSessionsGet.ts
@@ -38,7 +38,9 @@ export const userSessionsGet: APIFunction = async (_: APIInterface, options: API
                     const sessionResponse = await Session.getSessionInformation(response[i]);
 
                     if (sessionResponse !== undefined) {
-                        sessions[i] = sessionResponse;
+                        const accessTokenPayload = sessionResponse.customClaimsInAccessTokenPayload;
+                        delete sessionResponse.customClaimsInAccessTokenPayload;
+                        sessions[i] = { ...sessionResponse, accessTokenPayload };
                     }
 
                     res();